refactor: join페이지 폴더구조로 변경
