+++ frontend/src/pages/DogProfile.tsx
@@ -15,8 +15,8 @@ import Topbar from '@/components/common/Topbar';
 import FemaleIcon from '@/components/icon/FemaleIcon';
 import { MaleIcon } from '@/components/icon/MaleIcon';
 import { uploadImg, useDog } from '@/hooks/useDog';
-import { DogRegInfo } from '@/pages/Join';
-import { maxLengthCheck } from '@/pages/JoinStep/DogDetailInfo';
+import { DogRegInfo } from '@/pages/Join/index';
+import { maxLengthCheck } from '@/pages/Join/DogDetailInfo';
 import { useCropStore } from '@/store/cropStore';
 import { dataURLtoFile } from '@/utils/dataUrlToFile';
 import { secondsToTimeFormat } from '@/utils/time';
