+++ frontend/src/pages/Join/Agreements.tsx
