+++ frontend/src/pages/Join/DogDetailInfo.tsx
