+++ frontend/src/pages/Join/index.tsx
@@ -3,10 +3,10 @@ import { getStorage } from '@/utils/storage';
 import React, { useMemo, useRef, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import TopBack from '@/assets/icons/ic-top-back.svg';
-import Agreements from '@/pages/JoinStep/Agreements';
+import Agreements from '@/pages/Join/Agreements';
 import { Button } from '@/components/common/Button';
 import { Divider } from '@/components/common/Divider';
-import PetOwner from '@/pages/JoinStep/PetOwner';
+import PetOwner from '@/pages/Join/PetOwner';
 import Cancel from '@/assets/icons/ic-top-cancel.svg';
 import { storageKeys } from '@/constants';
 import { useAuth } from '@/hooks/useAuth';
@@ -17,8 +17,8 @@ import { dataURLtoFile } from '@/utils/dataUrlToFile';
 import { getUploadUrl } from '@/api/upload';
 import CancelRegModal from '@/components/CancelRegModal';
 import { useCropStore } from '@/store/cropStore';
-import DogBasicInfo, { DogBasicInfoProps } from '@/pages/JoinStep/DogBasicInfo';
-import DogDetailInfo, { DogDetailInfoProps } from '@/pages/JoinStep/DogDetailInfo';
+import DogBasicInfo, { DogBasicInfoProps } from '@/pages/Join/DogBasicInfo';
+import DogDetailInfo, { DogDetailInfoProps } from '@/pages/Join/DogDetailInfo';
 import { useSpinnerStore } from '@/store/spinnerStore';
 
 export interface DogRegInfo extends DogBasicInfoProps, DogDetailInfoProps {
